Refactor ModConfiguration
