+++ src/content/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(content_SRCS
 	${CMAKE_CURRENT_SOURCE_DIR}/content.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/mod_configuration.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/mods.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/subgames.cpp
 	PARENT_SCOPE
