+++ src/filesys.cpp
@@ -877,4 +877,3 @@ bool Rename(const std::string &from, const std::string &to)
 }
 
 } // namespace fs
-
